pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foxcode"
version = "0.1.5"
description = "FoxCode - 一个强大的 AI 终端编码助手"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "FoxCode Team", email = "foxcode@example.com" }
]
keywords = ["ai", "cli", "coding-assistant", "terminal", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "textual>=0.47.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tree-sitter>=0.21.0",
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "aiofiles>=23.0.0",
    "pathspec>=0.12.0",
    "playwright>=1.40.0",
    "numpy>=1.24.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
    "markdown-it-py>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
]
local = [
    "llama-cpp-python>=0.2.0",
]

[project.scripts]
foxcode = "foxcode.cli:main"

[project.urls]
Homepage = "https://github.com/foxcode/foxcode"
Documentation = "https://github.com/foxcode/foxcode#readme"
Repository = "https://github.com/foxcode/foxcode"
Issues = "https://github.com/foxcode/foxcode/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/foxcode"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
    "ARG001", # unused function argument
]

[tool.ruff.isort]
known-first-party = ["foxcode"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.coverage.run]
source = ["src/foxcode"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
